refactor: move get_ansys_process_from_port function to core module
