+++ tests/test_cli.py
@@ -32,7 +32,7 @@ import psutil
 import pytest
 
 import ansys.mapdl.core.cli.core as core_module
-from ansys.mapdl.core.cli.list_instances import get_ansys_process_from_port
+from ansys.mapdl.core.cli.core import get_ansys_process_from_port
 from ansys.mapdl.core.plotting import GraphicsBackend
 from conftest import VALID_PORTS, requires
